requirements.txt
matplotlib==3.2.1
numpy==1.18.4
speedtest-cli==2.1.2
